[Refactor] Review: 리뷰 docs 패키지 수정
