+++ review/src/main/java/com/haot/review/presentation/controller/ReviewController.java
@@ -7,10 +7,9 @@ import com.haot.review.application.dtos.res.ReviewGetResponse;
 import com.haot.review.application.service.ReviewService;
 import com.haot.review.common.response.ApiResponse;
 import com.haot.review.common.response.enums.SuccessCode;
+import com.haot.review.presentation.docs.ReviewControllerDocs;
 import com.haot.submodule.role.Role;
 import com.haot.submodule.role.RoleCheck;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
